Move validation to src/validation
